Step #46
What's a game without a story? Unless it doesn't need one that is ;)
We add a introduction page that will be shown for every chapter of the game. And a neat image of an Impala.

And I'm getting too close to my current progress, I hope I can continue without taking breaks.