Step #13
And yet again a bigger step. Of course we need lots of goodies from the killed enemies. We add items. Items are not sprites but a 2x2 block of characters. There's a new list of possible items with location added. The list also stores the original background behind the items.
To get an item spawned just walk beside on of the enemies (look in their direction) and keep fire pressed until it dies.

Note that the player cannot collect the items yet (that's up for the next step).