Step #9
What are enemies if they just sit put and don't move at all? Therefore we now add the sub routine ObjectControl. We provide a table with function pointers to the object's behaviour code (including the player).
ObjectControl takes the object type as index into the table and jumps to the object control. For now we have two enemy types (type is tied to behaviour), dumb moving up/down or left/right. For moving we reuse the previously created functions we already use for the player, namely ObjectMoveLeft/ObjectMoveRight etc.