Step #23
Highscore display is all nice and dandy, but we want to see our score and name in there! So this step adds name entry. For this we rely on the Kernal functions to read the pressed keys from the keyboard.

Adding to the score moving we find the spot for the name, shift all lower score names down and let the player enter his name.