Step #4
Now we take a bigger step: Moving the sprite with the joystick. Since we want to make a real game we also allow to move the sprite over to the right side (in other words we'll take care of the extended x bit).
Since I don't plan to allow sprites to go off screen there is no defined behaviour if you move the sprite off screen too far. It'll simply bounce back in once the coordinate wraps around.

This step shows:
-Joystick control
-Sprite extended x bit

The current state is step 36, satisfaction refinement. A new nasty and fast spider monster has appeared. Monsters are bounced back when shot and vanish in a puff of smoke when they die.