Step #11
From colliding to dying is a small step. Once the player collides with an enemy we kill him by removing the player object. A &quot;Press Fire to Restart&quot; message is displayed and a press on the button will revive the player object.

We add the function RemoveObject which simply removes the object from the SPRITE_ACTIVE table and disables its sprite. While we wait for the player to press the button all the rest of the game moves on.