Step #40
And onwards we go. This time we add spawn spots. These are a different kind of object which spawn a number of one specific enemy type. The level is clear once all enemies are killed and the spawn spots are empty.

In case you wonder about the .C64 file inside there: It's a project file for C64Studio, an .NET based IDE I've written for development and debugging alongside with mostly WinVICE.