Step #21
Now a technically more interesting step: We add a title screen with a bitmap logo. We're using a raster interrupt to show the bitmap logo in the top half of the screen and switch to text mode on the bottom half.

The bitmap data was created by a Lua script for my paint app. Then only the needed parts were cut out manually. We don't need the full 8k data for 8 lines.