Of course, just had my sister's wedding yesterday ;)
Step 15:

Now we start with a few game essentials: Progressing to the next level.

Not too complicated. We keep a counter of enemies alive (NUMBER_ENEMIES_ALIVE) which is initially set to 0. For every enemy object it's increased. And for every killed enemy it's decreased. Once the enemy count reaches 0 the level end delay kicks in (marked with the border flashing). Once the level end delay is complete the next level is displayed.
For now there's two simple levels with the last looping back to the first.