Step #3:
Quite similar to step 2 this time we set up sprites. Note that I used another selfmade tool (similar to SpritePad). The file J.SPR contains the sprites and is included into the source as binary. Since the C64 can't load files to locations after  we have to copy the charset to the target memory at  (and switch off the ROM therefore).
The position the sprites are now positioned is otherwise unused RAM that is &quot;hidden&quot; by ROM. The VIC however can read the RAM below. This makes that location quite convenient for sprites or charsets.

And another note: Since I'm lazy I built the sprite copy loop to always copy packages of 4 sprites (1 sprite comes as 63 + 1 byte, so 4 make nice 256 bytes).

The current step should display charset garbage and &quot;HELLO&quot; as before, but also display a elegantly handicrafted sprite.

Find attached the source and .prg file for your favourite emulator

The current state is step 34, gameplay refinement 8 (items). There are barely any visible changes so I skip the screenshot for this step.