Step #19
Nothing really new this time, just constant use of existing parts. A new enemy, some sort of mummy, which slowly walks back and forth. And takes a few more hits than the bats.

It's a new behaviour added to the table and some extra code that checks for gaps before walking.