Step 100
And thus this create a game series ends...

We fix the last few bugs (score was not properly reset on replay, music
was stuck on saving the scores). Now the game is complete and can be
enjoyed as it was meant to be!

Oh joy, Smila (the graphician) actually touched up the game even more for
a retail release. It has been out for a few weeks now, but if you didn't
know, go here:
Psytronik (for digital download, tape or disk) or to RGCD (cartridge with
nifty packaging and stickers).
Or simply marvel in all the nice games that are still being made for older
computers.

There you can actually buy a severely enhanced version (more two player
game modes, better graphic, gameplay enhancements) as either digital
download for emulators, or as a real tape, disk and finally as cartridge!

Thank you for your encouragements throughout, and keep on coding ;)